# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "description": "모든 유저 목록을 조회합니다.",
        "operationId": "UsersController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "모든 유저 목록 반환"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "모든 유저 목록 (관리자 전용)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "내 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 정보 조회",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 정보 수정 (이름, 이메일, 생년월일만)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "description": "ID를 사용하여 학생/유저를 조회합니다.",
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "description": "조회할 유저의 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "유저 정보 반환"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "유저 조회 (관리자 전용)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/api/consultations/devices": {
      "get": {
        "description": "조건에 맞는 단말기 목록과 계산된 가격(출고가, 공시지원금, 할부원금)을 반환합니다.",
        "operationId": "ConsultationsController_getDevices",
        "parameters": [
          {
            "name": "networkType",
            "required": true,
            "in": "query",
            "description": "유무선 구분",
            "schema": {
              "type": "string",
              "enum": [
                "WIRELESS",
                "WIRED"
              ]
            }
          },
          {
            "name": "carrier",
            "required": true,
            "in": "query",
            "description": "통신사",
            "schema": {
              "type": "string",
              "enum": [
                "SKT",
                "KT",
                "LGU"
              ]
            }
          },
          {
            "name": "searchType",
            "required": false,
            "in": "query",
            "description": "검색 조건 (기종명 또는 모델명)",
            "schema": {
              "type": "string",
              "enum": [
                "DEVICE_NAME",
                "MODEL_NAME"
              ]
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "검색어",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 단말기 목록을 반환했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "단말기 리스트 및 검색 API",
        "tags": [
          "Consultations (통합 유무선 상담 및 견적)"
        ]
      }
    },
    "/api/api/consultations/quotes/summary": {
      "get": {
        "description": "N+1 문제를 방지하기 위해 견적, 계약 상태, CRM 정보를 Aggregation하여 반환합니다.",
        "operationId": "ConsultationsController_getQuoteSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "BFF용 통합 견적 내역 조회 (CQRS Read Model)",
        "tags": [
          "Consultations (통합 유무선 상담 및 견적)"
        ]
      }
    },
    "/api/api/consultations/quotes": {
      "post": {
        "description": "선택한 단말기 정보를 바탕으로 현재 가격 스냅샷을 포함한 견적을 생성합니다.",
        "operationId": "ConsultationsController_createQuote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "성공적으로 견적을 생성했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "통합 견적 생성 API",
        "tags": [
          "Consultations (통합 유무선 상담 및 견적)"
        ]
      }
    }
  },
  "info": {
    "title": "전자계약 & CRM 플랫폼 API",
    "description": "전자계약 및 CRM 기능이 포함된 플랫폼 REST API 명세서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "reset-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-string"
          },
          "phoneNumber": {
            "type": "string",
            "example": "01012345678"
          },
          "name": {
            "type": "string",
            "example": "홍길동"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "STAFF"
            ],
            "example": "OWNER"
          },
          "referralCode": {
            "type": "string",
            "example": "ABCDEF"
          }
        },
        "required": [
          "id",
          "phoneNumber",
          "name",
          "role",
          "referralCode"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "홍길순",
            "description": "이름"
          },
          "phoneNumber": {
            "type": "string",
            "example": "01099998888",
            "description": "휴대폰 번호"
          },
          "password": {
            "type": "string",
            "description": "새 비밀번호 (최소 8자)"
          },
          "marketingAgreed": {
            "type": "boolean",
            "description": "마케팅 수신 동의"
          },
          "email": {
            "type": "string",
            "description": "이메일"
          },
          "birthDate": {
            "type": "string",
            "description": "생년월일 (YYYY-MM-DD)"
          }
        }
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password has been reset successfully."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "DeviceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "단말기 ID"
          },
          "networkType": {
            "type": "string",
            "description": "유무선 구분",
            "enum": [
              "WIRELESS",
              "WIRED"
            ]
          },
          "carrier": {
            "type": "string",
            "description": "통신사",
            "enum": [
              "SKT",
              "KT",
              "LGU"
            ]
          },
          "deviceName": {
            "type": "string",
            "description": "단말기 기종명"
          },
          "modelName": {
            "type": "string",
            "description": "모델명"
          },
          "retailPrice": {
            "type": "number",
            "description": "출고가"
          },
          "publicSubsidy": {
            "type": "number",
            "description": "공시지원금"
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "description": "출시일 (YYYY-MM-DD 형식)"
          },
          "principal": {
            "type": "number",
            "description": "할부원금 (출고가 - 공시지원금)"
          }
        },
        "required": [
          "id",
          "networkType",
          "carrier",
          "deviceName",
          "modelName",
          "retailPrice",
          "publicSubsidy",
          "releaseDate",
          "principal"
        ]
      },
      "QuoteSummaryDto": {
        "type": "object",
        "properties": {
          "quoteId": {
            "type": "string",
            "description": "견적 ID"
          },
          "quoteName": {
            "type": "string",
            "description": "견적명"
          },
          "deviceName": {
            "type": "string",
            "description": "기기명"
          },
          "principal": {
            "type": "number",
            "description": "단말기 할부원금"
          },
          "contractId": {
            "type": "string",
            "description": "전자계약 ID (매핑된 경우)"
          },
          "contractStatus": {
            "type": "string",
            "description": "전자서명 상태"
          },
          "crmCustomerId": {
            "type": "string",
            "description": "연관된 CRM 고객 ID"
          }
        },
        "required": [
          "quoteId",
          "quoteName",
          "deviceName",
          "principal"
        ]
      },
      "CreateQuoteDto": {
        "type": "object",
        "properties": {
          "networkType": {
            "type": "string",
            "description": "유무선 구분",
            "enum": [
              "WIRELESS",
              "WIRED"
            ]
          },
          "carrier": {
            "type": "string",
            "description": "통신사",
            "enum": [
              "SKT",
              "KT",
              "LGU"
            ]
          },
          "deviceId": {
            "type": "string",
            "description": "선택된 단말기 ID"
          }
        },
        "required": [
          "networkType",
          "carrier",
          "deviceId"
        ]
      },
      "Quote": {
        "type": "object",
        "properties": {}
      }
    }
  }
}